+++ packages/utils-dev-dependency/plaid/testData/index.ts
@@ -1,3 +1,4 @@
 export * from "./items";
 export * from "./eventsWebhooks";
 export * from "./accounts";
+export * from "./identity";