+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.16.0]
+## [1.17.0]
 
 ### Changed
 
 - Connecting screen now shows which aggregator is being used for the connection instead of the MX logo
+
+## [1.16.0]
+
+### Added
+
+- Plaid identity data endpoint for future plaid support
 
 ## [1.15.1]
 
@@ -18,9 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Flaky Plaid tests
 - Faster Plaid github actions by cacheing dependencies.
 
-### Disabled
+### Changed
 
-- Plaid adapter by default due to security risk
+- Disabled Plaid adapter by default due to security risk
 
 ## [1.15.0]